feat: :lipstick: Modificación de dos archivos .java
